requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
alembic==1.13.1
sqlalchemy==2.0.23
APScheduler==3.10.4
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
GitPython==3.1.40
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1

